Switch to new Python backup script
Turns out my old Bash backup script was in fact *not* handling commands
that returned non-zero. Using Python, all failing commands raise an
exception. For now, the whole script is aborted after the error is
reported, but in the future each error could be reported individually
while running the rest of the backup.